Implemented multiple views per room. Closes #23.
A room file can be split into views. Views are basically multiple rooms
packed together into the same room file, instead of each image /
direction you could be facing in the same conceptual room being a
separate room file. In other words, each area image is no longer a room,
it is a view inside the room.

Each room requires at least one view. Each view has a name, and is the
parent object for the data that comprises it. Rooms can reference a view
in another room to be loaded with "filename.json:viewname". If no view
is named, the "default" view of the room will be loaded.